Add files via upload
This code is a script that accepts a valid Wikipedia link and a valid integer between 1 to 20 (n) as input. It then scrapes the link provided for all wiki links embedded in the page and stores them in a list. The script then recursively calls itself to get all the wiki links at the given level and stores them in the same list. It optimizes the code not to visit any links you've already visited. Finally, it writes the results (all found links, total count, unique count) to a CSV/JSON file.

The code uses the following methods and functions: 

1) input() - This is a built-in function that reads a line of text from the user. 

2) re.match() - This is a method from the re module that attempts to match a regular expression pattern to a string. 

3) requests.get() - This is a method from the requests module that is used to make a GET request to a specified URL. 

4) re.findall() - This is a method from the re module that finds all substrings in a string that match a given regular expression pattern.

5) open() - This is a built-in function that opens a file and returns a file object. 

6) csv.writer() - This is a method from the csv module that returns a writer object that can be used to write to a CSV file. 

7) json.dump() - This is a method from the json module that is used to write a Python object to a JSON file.